feat: resume the change stream upon reconnection
Upon reconnection to the MongoDB server, the client will now try to
resume the stream at the last offset it has processed.

If the MongoDB client is disconnected for too long and its token is no
longer valid, then the Socket.IO clients connected to this server may
miss some packets (which was the previous behavior).